moved blackbox greedy probs calc to a separate file, small bugfix
