+++ scripts/polygraph_eval
@@ -299,7 +299,7 @@ def get_blackbox_model(args):
         return BlackboxModel.from_openai(
             openai_api_key=openai_api_key,
             model_path=args.model.path,
-            supports_logprobs=args.supports_logprobs,
+            supports_logprobs=args.model.supports_logprobs,
         )
     elif provider == "huggingface":
         hf_api_key = os.environ.get("HUGGINGFACE_API_KEY")
