+++ src/lm_polygraph/stat_calculators/__init__.py
@@ -2,6 +2,8 @@ from .stat_calculator import StatCalculator
 from .initial_state import InitialStateCalculator
 from .greedy_probs import (
     GreedyProbsCalculator,
+)
+from .greedy_probs_blackbox import (
     BlackboxGreedyTextsCalculator,
 )
 from .greedy_lm_probs import GreedyLMProbsCalculator
