+++ src/lm_polygraph/stat_calculators/greedy_probs.py
@@ -5,153 +5,7 @@ from typing import Dict, List, Tuple
 
 from .embeddings import get_embeddings_from_output
 from .stat_calculator import StatCalculator
-from lm_polygraph.utils.model import WhiteboxModel, BlackboxModel
-
-
-class BlackboxGreedyTextsCalculator(StatCalculator):
-    """
-    A calculator that handles both blackbox and greybox model generation.
-    If the model supports logprobs, it provides full greybox functionality (texts, logprobs, etc.).
-    Otherwise, it falls back to basic blackbox functionality (texts only).
-    """
-
-    @staticmethod
-    def meta_info() -> Tuple[List[str], List[str]]:
-        """
-        Returns the statistics and dependencies for the calculator.
-        """
-        return [
-            "greedy_texts",
-            "greedy_log_probs",
-            "greedy_log_likelihoods",
-            "greedy_tokens",
-        ], []
-
-    def __init__(
-        self,
-        top_logprobs: int = 5,
-    ):
-        super().__init__()
-        self.top_logprobs = top_logprobs
-
-    def __call__(
-        self,
-        dependencies: Dict[str, np.array],
-        texts: List[str],
-        model: BlackboxModel,
-        max_new_tokens: int = 100,
-    ) -> Dict[str, np.ndarray]:
-        """
-        Calculates generation texts and log probabilities for the model.
-        Provides full greybox functionality if the model supports logprobs,
-        otherwise falls back to basic blackbox functionality.
-
-        Parameters:
-            dependencies (Dict[str, np.ndarray]): input statistics, can be empty (not used).
-            texts (List[str]): Input texts batch used for model generation.
-            model (BlackboxModel): Model used for generation.
-            max_new_tokens (int): Maximum number of new tokens at model generation. Default: 100.
-        Returns:
-            Dict[str, np.ndarray]: dictionary with model generations and, if supported, probability data:
-                - 'greedy_texts' (List[str]): model generations corresponding to the inputs,
-                - 'greedy_log_probs' (List[List[np.array]]): logits for the top k tokens (greybox only),
-                - 'greedy_log_likelihoods' (List[List[float]]): log-probabilities of generated tokens (greybox only),
-                - 'greedy_tokens' (List[List[str]]): tokens of the generated text (greybox only).
-        """
-        if model.supports_logprobs:
-            # Greybox path: generate with logprobs
-            sequences = model.generate_texts(
-                input_texts=texts,
-                max_new_tokens=max_new_tokens,
-                n=1,
-                output_scores=True,
-                top_logprobs=self.top_logprobs,
-            )
-        else:
-            # Blackbox path: generate just the text
-            sequences = model.generate_texts(
-                input_texts=texts,
-                max_new_tokens=max_new_tokens,
-                n=1,
-            )
-
-        # Process the results
-        greedy_texts = sequences
-        greedy_log_probs = []
-        greedy_log_likelihoods = []
-        greedy_tokens = []
-
-        # Extract logprobs and tokens from the model's stored data if available
-        if model.supports_logprobs and hasattr(model, "logprobs") and model.logprobs:
-            for i, logprob_data in enumerate(model.logprobs):
-                if hasattr(logprob_data, "content"):
-                    # Extract tokens
-                    tokens = [item.token for item in logprob_data.content]
-                    greedy_tokens.append(tokens)
-
-                    # Extract log probabilities for this generation
-                    log_probs_list = []
-                    log_likelihoods = []
-
-                    for token_logprobs in logprob_data.content:
-                        # Get the top logprobs for this token position from OpenAI API
-                        token_logprob_dict = {}
-                        for top_logprob in getattr(token_logprobs, "top_logprobs", []):
-                            token_logprob_dict[top_logprob.token] = top_logprob.logprob
-
-                        # Create a sparse representation of the logprobs distribution
-                        sparse_logprobs = np.ones(
-                            model.model_path_vocab_size
-                            if hasattr(model, "model_path_vocab_size")
-                            else 50000
-                        ) * -float("inf")
-
-                        # Map token strings to positions in the sparse array
-                        for token_str, logprob in token_logprob_dict.items():
-                            token_idx = hash(token_str) % len(sparse_logprobs)
-                            sparse_logprobs[token_idx] = logprob
-
-                        log_probs_list.append(sparse_logprobs)
-
-                        # Extract the log probability of the chosen token
-                        chosen_logprob = token_logprobs.logprob
-                        log_likelihoods.append(chosen_logprob)
-
-                    greedy_log_probs.append(log_probs_list)
-                    greedy_log_likelihoods.append(log_likelihoods)
-
-            # Ensure all outputs have the same length for greybox case
-            while len(greedy_tokens) < len(greedy_texts):
-                # If we're missing token data, add placeholder empty lists
-                greedy_tokens.append([])
-                greedy_log_probs.append([])
-                greedy_log_likelihoods.append([])
-
-            result = {
-                "greedy_texts": greedy_texts,
-                "greedy_log_probs": greedy_log_probs,
-                "greedy_log_likelihoods": greedy_log_likelihoods,
-                "greedy_tokens": greedy_tokens,
-            }
-        else:
-            # For blackbox models, only return the generated texts
-            result = {"greedy_texts": greedy_texts}
-
-            # If user explicitly requests logprobs functionality but model doesn't support it, raise error
-            if any(
-                key in dependencies
-                for key in [
-                    "greedy_log_probs",
-                    "greedy_log_likelihoods",
-                    "greedy_tokens",
-                ]
-            ):
-                raise ValueError(
-                    "Model must support logprobs for retrieving probability information. "
-                    "The current model only supports text generation."
-                )
-
-        return result
+from lm_polygraph.utils.model import WhiteboxModel
 
 
 class GreedyProbsCalculator(StatCalculator):
